Move things around + cargo fmt
