+++ tests/validate_did.rs
@@ -1,4 +1,4 @@
-use did_common::{Did};
+use did_common::did::Did;
 
 #[test]
 fn validate_generic_did() {
@@ -7,10 +7,7 @@ fn validate_generic_did() {
 
 #[test]
 fn validate_generic_did_with_fragment() {
-    assert_eq!(
-        Did::is_valid("did:example:123456789abcdefghi#keys-1"),
-        true
-    )
+    assert_eq!(Did::is_valid("did:example:123456789abcdefghi#keys-1"), true)
 }
 
 #[test]
@@ -25,7 +22,10 @@ fn validate_did_with_invalid_method_name() {
 
 #[test]
 fn validate_ethr_did() {
-   assert_eq!(Did::is_valid("did:ethr:0xf3beac30c498d9e26865f34fcaa57dbb935b0d74"), true);
+    assert_eq!(
+        Did::is_valid("did:ethr:0xf3beac30c498d9e26865f34fcaa57dbb935b0d74"),
+        true
+    );
 }
 
 #[test]
@@ -35,7 +35,10 @@ fn validate_did_with_generic_param() {
 
 #[test]
 fn validate_did_with_multiple_params() {
-    assert_eq!(Did::is_valid("did:example:1234;service=agent;example:foo:bar=baz"), true);
+    assert_eq!(
+        Did::is_valid("did:example:1234;service=agent;example:foo:bar=baz"),
+        true
+    );
 }
 
 #[test]
